+++ src/server/http/Request.js
@@ -1,9 +1,22 @@
-import Headers from './Headers'
+import HttpMessage from './HttpMessage'
 
-export default class Request {
+export default class Request extends HttpMessage {
   constructor (method, url, headers) {
+    super(headers)
+
     this.method = method
     this.url = url
-    this.headers = headers || new Headers()
+  }
+
+  get _protected () {
+    return super._protected.concat('method', 'url')
+  }
+
+  _copy () {
+    return new Request(
+      this.method,
+      this.url,
+      this.headers
+    )
   }
 }